requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
weaviate-client>=4.6.0
sentence-transformers
numpy
tqdm
httpx
python-multipart